Move cookie modal to footer
