--- frontend/pages/privacy.vue
@@ -1,17 +1,0 @@
-<template>
-  <BasePageLayout>
-    <BaseContainer>
-      <h1>PRIVACY & COOKIES</h1>
-    </BaseContainer>
-  </BasePageLayout>
-</template>
-
-<script lang="ts">
-import { defineComponent } from '@nuxtjs/composition-api'
-
-export default defineComponent({
-  setup() {},
-})
-</script>
-
-<style lang="scss" scoped></style>
